requirements.txt
joblib==1.3.2
matplotlib==3.8.2
mysql-connector-python==8.2.0
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
plotly==5.18.0
pymysql==1.1.0
seaborn==0.13.0
scikit-learn==1.3.2
sequential==1.0.0
Sphinx==7.2.6
streamlit==1.28.2
tensorflow==2.15.0
toml==0.10.2